requirements.txt
django-rest-auth
django-allauth
Django==2.1.5
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
